backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
httpx==0.28.1
sse-starlette==2.2.1
pydantic==2.10.4
python-dotenv==1.0.1
youtube-transcript-api>=1.0.0
sentence-transformers>=2.2.0
numpy>=1.24.0
psycopg2-binary>=2.9.0
pgvector>=0.2.0
